Added cart with 2 endpoints

+++ Zadanie4/models/Product.go
@@ -4,6 +4,6 @@ import "gorm.io/gorm"
 
 type Product struct {
 	gorm.Model
-	Name  string `json:"name"`
-	Price uint   `json:"price"`
+	Name  string  `json:"name"`
+	Price float64 `json:"price"`
 }
